Two fixes provided by Trira.
git-svn-id: svn://svn.code.sf.net/p/oblivionworks/code/Programs/Wrye%20Bash@2378 c80116c5-1073-4541-b2a9-63b9fd661f84

+++ Mopy/bash/bosh.py
@@ -5644,6 +5644,7 @@ class ModInfos(FileInfos):
             #--Select in plugins
             if fileName not in boss.ActivePlugins:
                 boss.ActivePlugins.append(fileName)
+                self.refreshFile(fileName)
         finally:
             if doSave:
                 self.refreshInfoLists()
@@ -5747,8 +5748,9 @@ class ModInfos(FileInfos):
 
     def delete(self,fileName,doRefresh=True):
         """Deletes member file."""
-        self.unselect(fileName)
-        FileInfos.delete(self,fileName,doRefresh)
+        if fileName.s != u'Oblivion.esm' and fileName.s != u'Nehrim.esm' and fileName.s != u'Skyrim.esm' and fileName.s != u'Update.esm':
+            self.unselect(fileName)
+            FileInfos.delete(self,fileName,doRefresh)
 
     def move(self,fileName,destDir,doRefresh=True):
         """Moves member file to destDir."""
